refactor: split god classes exceeding 1000 lines
Split status.ts (1803 lines) into status/ submodule:
- status/types.ts (194 lines): all interfaces and constants
- status/renderers.ts (382 lines): terminal output rendering
- status/data.ts (937 lines): data loading functions
- status/index.ts (386 lines): command orchestration

Split researchCoordinator.ts (1367 lines) into companion modules:
- researchDetection.ts (529 lines): unknown detection pipeline
- researchPersistence.ts (165 lines): task I/O and caching
- researchCoordinator.ts (730 lines): coordinator class and types

No functionality changes — pure structural refactor.
All existing imports remain backward-compatible.

Closes CDMCH-97
